refactor(screens): screens 디렉토리 구조를 기능별로 재정리
 community 관련 컴포넌트를 community-stock으로 이동
 post 관련 컴포넌트를 edit-post, post-detail, write-post로 분리
 mypage 관련 컴포넌트를 mycomment, mypost, myprofile로 분리
 onboarding, search-stock, shortform 컴포넌트 추가
 app 라우트의 layout.tsx, page.tsx 파일들 업데이트
