--- apps/web/app/onboarding/KakaoLoginButton.tsx
@@ -1,1 +1,0 @@
-export { default } from '@/screens/onboarding/ui/KakaoLoginButton';
