--- apps/web/app/onboarding/Logo.tsx
@@ -1,1 +1,0 @@
-export { default } from '@/screens/onboarding/ui/Logo';
