+++ apps/web/src/screens/community-stock/ui/CommunityStockLayout.tsx
