+++ apps/web/src/screens/community-stock/ui/CompanyInfoBox.tsx
