+++ apps/web/src/screens/post-detail/ui/PostDetailLayout.tsx
