+++ apps/web/src/screens/post-detail/ui/PostSection.tsx
