+++ apps/web/src/screens/write-post/ui/WritePostLayout.tsx
